Final paper evaluation code. - Refactor Evaluation - Reinstate ELMo - Add bootstrapping - Add Hugging Face BERT - Track test set during pre-training

+++ modules/bsg/bsg_main.py
@@ -1,8 +1,9 @@
-import pickle
+import csv
 import os
+import pickle
 from shutil import rmtree
 import sys
-from time import sleep
+from time import sleep, time
 
 import argparse
 import numpy as np
@@ -10,13 +11,18 @@ import torch
 from tqdm import tqdm
 
 home_dir = os.path.expanduser('~/LMC/')
+sys.path.insert(0, os.path.join(home_dir, 'acronyms'))
+sys.path.insert(0, os.path.join(home_dir, 'acronyms', 'modules'))
 sys.path.insert(0, os.path.join(home_dir, 'preprocess'))
 sys.path.insert(0, os.path.join(home_dir, 'utils'))
+from acronym_utils import load_mimic, load_casi, load_columbia
+from bsg_acronym_expander import BSGAcronymExpander
 from bsg_batcher import BSGBatchLoader
 from bsg_model import BSG
-from bsg_utils import save_checkpoint
+from bsg_utils import restore_model, save_checkpoint
 from compute_sections import enumerate_metadata_ids_multi_bsg
-from model_utils import get_git_revision_hash, render_args
+from evaluate import run_evaluation
+from model_utils import block_print, enable_print, get_git_revision_hash, render_args, render_num_params
 
 
 if __name__ == '__main__':
@@ -28,7 +34,7 @@ if __name__ == '__main__':
 
     # Training Hyperparameters
     parser.add_argument('--batch_size', default=1024, type=int)
-    parser.add_argument('--epochs', default=4, type=int)
+    parser.add_argument('--epochs', default=5, type=int)
     parser.add_argument('--lr', default=0.001, type=float)
     parser.add_argument('--window', default=10, type=int)
 
@@ -39,6 +45,7 @@ if __name__ == '__main__':
     parser.add_argument('--multi_weights', default='0.7,0.2,0.1')
     parser.add_argument('--mask_p', default=None, type=float, help=(
         'Mask Encoder tokens with probability mask_p if a float.  Otherwise, default is no masking.'))
+    parser.add_argument('-restore', default=False, action='store_true')
 
     args = parser.parse_args()
     args.git_hash = get_git_revision_hash()
@@ -94,22 +101,44 @@ if __name__ == '__main__':
     batcher = BSGBatchLoader(len(ids), all_metadata_pos_idxs, batch_size=args.batch_size)
 
     # Instantiate PyTorch BSG Model
-    model = BSG(args, vocab.size()).to(device_str)
+    if args.restore:
+        print('Restoring from latest checkpoint...')
+        epoch_shift = 7  # TODO determine from checkpoints
+        _, model, _, optimizer_state = restore_model(args.experiment)
+        model = model.to(device_str)
+    else:
+        epoch_shift = 0
+        model = BSG(args, vocab.size()).to(device_str)
+        optimizer_state = None
+    render_num_params(model, vocab.size())
 
     # Instantiate Adam optimizer
     trainable_params = filter(lambda x: x.requires_grad, model.parameters())
     optimizer = torch.optim.Adam(trainable_params, lr=args.lr)
+    if optimizer_state is not None:
+        print('Loading previous optimizer state')
+        optimizer.load_state_dict(optimizer_state)
 
     # Create model experiments directory or clear if it already exists
     weights_dir = os.path.join(home_dir, 'weights', 'bsg', args.experiment)
-    if os.path.exists(weights_dir):
-        print('Clearing out previous weights in {}'.format(weights_dir))
-        rmtree(weights_dir)
-    os.mkdir(weights_dir)
+    if not args.restore:
+        if os.path.exists(weights_dir):
+            print('Clearing out previous weights in {}'.format(weights_dir))
+            rmtree(weights_dir)
+        os.mkdir(weights_dir)
+
+    metric_cols = ['examples', 'lm_kl', 'lm_recon', 'epoch', 'hours', 'dataset', 'log_loss', 'accuracy', 'macro_f1',
+                   'weighted_f1']
+    metrics_file = open(os.path.join(weights_dir, 'metrics.csv'), mode='a')
+    metrics_writer = csv.writer(metrics_file, delimiter=',', quotechar='"', quoting=csv.QUOTE_MINIMAL)
+    metrics_writer.writerow(metric_cols)
+    metrics_file.flush()
+
+    start_time = time()
 
     # Make sure it's calculating gradients
     model.train()  # just sets .requires_grad = True
-    for epoch in range(1, args.epochs + 1):
+    for epoch in range(1 + epoch_shift, args.epochs + epoch_shift + 1):
         sleep(0.1)  # Make sure logging is synchronous with tqdm progress bar
         print('Starting Epoch={}'.format(epoch))
         batcher.reset()
@@ -133,6 +162,8 @@ if __name__ == '__main__':
 
             checkpoint_interval = 10000
             if (i + 1) % checkpoint_interval == 0:
+                duration_in_hours = (time() - start_time) / (60. * 60.)
+                full_example_ct = (((epoch - 1) * float(num_batches)) + i + 1) * args.batch_size
                 print('Saving Checkpoint at Batch={}'.format(i + 1))
                 d = float(i + 1)
                 # Serializing everything from model weights and optimizer state, to to loss function and arguments
@@ -143,6 +174,30 @@ if __name__ == '__main__':
                 print(losses_dict)
                 checkpoint_fp = os.path.join(weights_dir, 'checkpoint_{}.pth'.format(epoch))
                 save_checkpoint(args, model, optimizer, vocab, losses_dict, checkpoint_fp=checkpoint_fp)
+
+                experiments = [(load_casi, 'casi'), (load_mimic, 'mimic'), (load_columbia, 'columbia')]
+                for loader, dataset in experiments:
+                    args.lm_type = 'bsg'
+                    args.lm_experiment = args.experiment
+                    args.ckpt = None
+                    args.device = device_str
+                    prev_epoch_ct = args.epochs
+                    args.epochs = 0
+                    block_print()
+                    metrics = run_evaluation(args, BSGAcronymExpander, loader, restore_model, train_frac=0)
+                    enable_print()
+                    args.epochs = prev_epoch_ct
+                    metrics['dataset'] = dataset
+                    metrics['hours'] = duration_in_hours
+                    metrics['examples'] = full_example_ct
+                    metrics['epoch'] = epoch
+                    metrics['lm_recon'] = losses_dict['losses']['recon']
+                    metrics['lm_kl'] = losses_dict['losses']['kl']
+                    row = [metrics[col] for col in metric_cols]
+                    metrics_writer.writerow(row)
+                    print(metric_cols)
+                    print(row)
+                    metrics_file.flush()
 
         epoch_joint_loss /= float(batcher.num_batches())
         epoch_kl_loss /= float(batcher.num_batches())
@@ -156,3 +211,4 @@ if __name__ == '__main__':
         losses_dict = {'losses': {'joint': epoch_joint_loss, 'kl': epoch_kl_loss, 'recon': epoch_recon_loss}}
         checkpoint_fp = os.path.join(weights_dir, 'checkpoint_{}.pth'.format(epoch))
         save_checkpoint(args, model, optimizer, vocab, losses_dict, checkpoint_fp=checkpoint_fp)
+    metrics_file.close()
